separate swell mainnet/testnet logos
