+++ src/index.tsx
@@ -27,7 +27,8 @@ import baseSepoliaLogo from './assets/images/networks/base_network_logo_with_mar
 import blastSepoliaLogo from './assets/images/networks/blast_sepolia_logo.webp';
 import plumeSepoliaLogo from './assets/images/networks/plume_mainnet_logo_small.webp';
 import scrollSepoliaLogo from './assets/images/networks/scroll_sepolia_logo.webp';
-import swellSepoliaLogo from './assets/images/networks/swell_network_logo_with_margin.webp';
+import swellLogo from './assets/images/networks/swell_network_logo_with_margin.webp';
+import swellSepoliaLogo from './assets/images/networks/swell_sepolia.webp';
 
 // /* Perform a single forcible reload when the page first loads. Without this, there
 //  * are issues with Metamask and Chrome preloading. This shortcircuits preloading, at the
@@ -85,7 +86,7 @@ const modal = createWeb3Modal({
         534352: scrollLogo,
         11155111: sepoliaLogo,
         98864: plumeSepoliaLogo,
-        1923: swellSepoliaLogo,
+        1923: swellLogo,
         1924: swellSepoliaLogo,
         84532: baseSepoliaLogo,
     },
